Fix some buggy text encoding/decoding steps, mostly by removing them.
There were some rendering steps that would defensively encode and decode
Unicode in various ways. These steps were occasionally buggy, but furthermore
they seem to be unnecessary.

Python 2.7 (with unicode_literals) and 3.x are both happy to work with text
that's Unicode the entire time, as long as we use "codecs.open" instead of
"open" for Python 2.7's benefit. So my change largely consists of switching
to codecs.open and removing the re-encoding steps.

This has the following benefits:

- Fixes a bug where literals in rules would get mojibaked -- for example,
  "t?" would become "t??", due to some escaping step that confused
  Latin-1 and UTF-8
- Trace output is now legible on Python 3 (one step was encoding its argument
  and never decoding it, causing it to show up as a byte literal on Py3)
- Fewer function calls should lead to a slight gain in efficiency