Changes to support nested left recursion
Previously the code had two variables which tracked the
current left recursion being evaluated. If a second left
recursion was met which wasn't a left-decendant of the
first rule (i.e. a token was matched before the second
recursive rule) then the evaluation of the second rule
corrupted the state tracking for the first rule.

The fix is to turn the tracking variables into stacks.